requirements.txt
google-cloud-storage==2.13.0
python-dotenv==1.0.0